pyproject.toml
[tool.poetry]
name = "yt-python-drops"
version = "0.1.0"
description = ""
authors = ["André Lima"]
readme = "README.md"
packages = [{include = "yt_python_drops"}]

[tool.poetry.dependencies]
python = "^3.8"
memory-profiler = "^0.61.0"
matplotlib = "^3.7.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
